refactor(cli): extract StatusCommand to separate file to fix SwiftLint file length violation
- Move StatusCommand class from Commands.swift to StatusCommand.swift
- Reduce Commands.swift from 937 to 690 lines (below 500 line limit)
- Maintain all functionality with proper imports and logger setup
- Fix remaining CI SwiftLint failure for clean PR status
