+++ .spec-workflow/specs/system-extension-integration/tasks.md
@@ -178,7 +178,7 @@ The System Extension Integration implementation follows a phased approach that b
   - _Leverage: existing CI pipeline and test infrastructure_
   - _Requirements: All_
 
-- [-] 17. Create pull request and validate CI
+- [x] 17. Create pull request and validate CI
   - Commit all changes with proper commit messages as specified in tasks above
   - Push branch: `git push -u origin feature/system-extension`
   - Create pull request with comprehensive description
